pyproject.toml
[project]
name = "rule-score-auto-adjust"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair[all]>=5.5.0",
    "diskcache>=5.6.3",
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.23",
    "mlflow>=3.1.0",
    "networkx[default]>=3.5",
    "perfect>=0.0.1",
    "polars>=1.30.0",
    "ray[data,default]>=2.46.0",
    "scikit-learn>=1.6.1",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "pathvalidate",
    "pyzipper>=0.3.6",
    "prefect>=3.6.10",
    "quark-engine",
]

[tool.uv]
environments = ["sys_platform == 'linux'"]

[tool.uv.sources]
quark-engine = { path = "../quark-engine" }
